refactor: setting api hook을 shared/ 밑으로 이동 및 경로 수정
