+++ src/pages/login/hooks/useGuestLogin.ts
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 
 import { postGuestLogin } from '@/pages/login/apis/authApi';
 import { setAuthTokens, setLoginType } from '@/pages/login/utils/tokenStorage';
-import { getUsers } from '@/pages/setting/apis/getUsers';
+import { getUsers } from '@/shared/apis/getUsers';
 
 export const useGuestLogin = () => {
   const navigate = useNavigate();
