+++ src/shared/hooks/useGetUsers.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getUsers } from '@/pages/setting/apis/getUsers';
+import { getUsers } from '@/shared/apis/getUsers';
 
 // 내 정보를 조회하는 query 훅.
 export const useGetUsers = () => {
